// frontend/src/shared/lib/api/model/componentRelationship.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * prot-cellar
 * OpenAPI spec version: 0.1.0
 */

export type ComponentRelationship = typeof ComponentRelationship[keyof typeof ComponentRelationship];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ComponentRelationship = {
  single_protein: 'single_protein',
  protein_subunit: 'protein_subunit',
  family_member: 'family_member',
  interacting_protein: 'interacting_protein',
} as const;